Add files via upload
Just created a currency converter with API calls! It was an amazing and fun learning journey. I also fixed the error caused by the API link change in the Apna College course. Check out my GitHub for the code and see the highlighted part in the screenshot. check out my twitter